S302T300-82 #time 1h 40m ?implement country data now uses scheduler and saves to file
Instead of storing only locally, the country data is stored in a file.

This means that in testing, a custom file can be passed to the validator so that tests do not rely on having an internet connection.

Replaced FetchRequestException with IllegalStateException; before some methods can be called data must first be loaded