S02T300-260 #time 1h ?implement user profile grid layout for mobile
Needed to fiddle with SASS imports to get breakpoints working - can't use CSS variables inside media queries so had to do it in SCSS, but had issues with importing bootstrap's _variables.scss file because of undefined functions or varaibles, even when _functions.scss was imported.

Ended up creating a new file, grid-breakpoints which is the source of truth for breakpoint widths, and imported it into main and the Details.vue page.