S302T300-260 #time 30m ?fix height bug in mobile
User details table could overflow, so tried to make it scrollable vertically in case it got too large. Got it scrolling horizontally, which is a plus.

Real issue turned out to be a previous commit where I added a class to a rule, but didn't cancel it for mobile users.

Also spent a while on a wild goose chase trying to figure out why there was no margin on the right hand side of the screen - the issue is that the Chrome dev tool's mobile simulator's grabby thingy for resizing the window (only in responsive mode!) is hiding a few pixels on the right edge. Ahhhhhhhhh!