S302T300-260 #time 20m ?implement scroll user details table
Necessary as I can't find a way to make the businesses list have
a maximum height, but grow past that height if the user details
container is taller than that. Current solution is to have the
both containers have a maximum height, but for certain screen widths
this can be exceeded. In these cases, the users details list can also
scroll