S302T300-204 #time 15m ?fix minor search inconsistency, CSS changes
If user was added as admin, when you click on the search field the user you just added would not be disabled. Changed to computed property to fix this bug.

Table icons now work, and search field is now wider

Took a few minutes to delete this commit after accidentally committing some extra files. Fun with Git