S302T300-204 #time 30m ?implement better error handling to Edit Admin page
addAdmin now handles errors.

Took a long time to figure out I was using `Api.handle401.bind` instead of `.call`, which caused the if statement to evaluate to true and hence the error message to not appear below the search field.

Remove admin logic added; just waiting for backend, the other code is mostly handled, except how to show the error message to the user. Not sure how to do that yet - also an issue with user search calls