?fix failing Jest tests
Time for this has been logged in another task on a different branch.

setup.js adds Vuetify to Vue.

To test files with Vuetify components, simply add
```
vuetify: new Vuetify() // (or define vuetify as a variable before hand; you don't need to create a new Vuetify instance for every test)
```
as a `mount` or `shallowMount` parameter.