S302T300-232 #time 25m ?test listing service delete method
When intentionally making the test fail to see if it was caught, found another bug that I fixed - an infinite loop would occur when Mockito tried to helpfully print out the difference between the expected and received objects. As objects referenced each other, Lombok's toString method would try and print out all properties, leading to an infinite loop that results in a stack overflow. Fun stuff.