FullInventoryItemDto returns business
Required for frontend - on listings search page, currency is required. WIthout this information, an API request to the business endpoint would be required (to get business country) for every listing returned.

BasicInventoryItem and FullInventoryItemDto had all the same fields, so I deleted BasicInventoryItemDto and replaced references to it with FullInventoryItemDto.