S302T300-375 #time 20m ?implement show full text in allergy info chips
After feedback from MR, realized it doesn't work well on mobile as clicking the chip sends you to the listing page.

Best solution seems to be to be just show the full text: it doesn't take too much extra space. Also simplifies CSS for the palm-oil free chip.

Allergy chip info was moved to constants.js so that the colors could be used in other components