S302T300-409 #time 15m ?implement add nutrient info to data.sql
As far as I can see there's no way to mock the `findAll` method and there are no methods to view the contents of a `Specification` object to check that it has the right predicates. Hence, using the test database seems to be the only way of testing the new filters.