S302T300-363 #time 35m ?test filtering by nutrition levels
Not an ideal situation with data.sql - I was missing product with ID 5002 and the listing IDs don't really match up with the product IDs, and although I could update them to match this will cause other tests to break.

The only alternative I can think of is to nuke the database after each test and add the users, businesses, products, inventory items and listings we need for each test, but that would take several hours to program and isn't worth it.