Add pigeon logo to navbar, get favicon working
Navbar pigeon logo added. Also switched from a link with JS to a `router-link` for a nicer experience.

Favicon: index.html must be in public/index.html. As it was in assets/index.html, it was not being used at all so the favicon was not working (in addition, BASE_URL was not defined - our app use VUE_APP_BASE_URL)

In addition, the following changes were made:

- noscript text was enabled
- Default title (defined in vue.config.js) was added to the app
- A redundant `div id="app"` was removed from App.vue - this is defined in index.html so there were two divs with the same ID